Fix sorting ambiguity in fragmenter
We changed the behavior when we removed qsort_r. Previously we never
allowed "equal" entries since this made it up to the whim of the
implementation regarding whether or not to place the entry before or
after the next one. Now (as in before) we use the actual pointer
comparison (they are part of an array)